==> notebooks/01_EXCEL_CSV.ipynb <==
import os
import pandas as pd
import os
import pandas as pd
import re
from openpyxl import load_workbook
import sys

# Paths
input_folder = r"D:\TRAINING MODEL\data\excel_folder"
output_folder = r"D:\TRAINING MODEL\data\csv_folder"
log_file = r"D:\TRAINING MODEL\logs\converted_files.txt"

# Create output folder if it doesn't exist
os.makedirs(output_folder, exist_ok=True)
os.makedirs(os.path.dirname(log_file), exist_ok=True)


if os.path.exists(log_file):
    with open(log_file, "r") as f:
        converted_files = f.read().splitlines()
else:
    converted_files = []

# Loop through all Excel files
for file_name in os.listdir(input_folder):
    if file_name.endswith(".xlsx") and file_name not in converted_files:
        file_path = os.path.join(input_folder, file_name)
        df = pd.read_excel(file_path)

        # Generate CSV file name
        csv_file_name = file_name.replace(".xlsx", ".csv")
        csv_path = os.path.join(output_folder, csv_file_name)

        # Save to CSV
        df.to_csv(csv_path, index=False)
        print(f"Converted {file_name} to {csv_file_name}")

        # Add to log
        with open(log_file, "a") as f:
            f.write(file_name + "\n")

print("Processing complete. Already converted files were skipped.")
